requirements.txt
mariadb==1.1.12
mysql-connector-python==9.3.0
packaging==25.0
python-dotenv==1.1.0
tabulate==0.9.0
